backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.7.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.21.0
